Make all system specs for current scenarios on course choices to pass
