+++ spec/forms/candidate_interface/continuous_applications/course_selection_wizard_spec.rb
