+++ spec/support/test_helpers/candidate_helper.rb
@@ -622,4 +622,11 @@ module CandidateHelper
   def current_candidate
     @current_candidate ||= create(:candidate)
   end
+
+  def then_i_should_be_on_the_application_choice_review_page
+    expect(application_choice).to be_present
+    expect(page).to have_current_path(
+      candidate_interface_continuous_applications_course_review_path(application_choice_id: application_choice.id),
+    )
+  end
 end
